% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_error_catch.R
\name{api_error_catch}
\alias{api_error_catch}
\title{Handle and parse errors from the API calls}
\usage{
api_error_catch(bodylist, call = rlang::caller_env(), .errorhandling = "stop")
}
\arguments{
\item{bodylist}{list returned by API}

\item{call}{see \code{\link[rlang:abort]{rlang::abort()}}; passes calling environment for clearer
messages}

\item{.errorhandling}{as in \code{\link[foreach:foreach]{foreach::foreach()}}- passed in here to handle
simply \code{abort}ing vs doing something with the error and continuing}
}
\value{
depends on presence of errors and value of \code{.errorhandling}
}
\description{
Handle and parse errors from the API calls
}
\keyword{internal}
