% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_times.R
\name{extract_timezone}
\alias{extract_timezone}
\title{Get timezones from various formats}
\usage{
extract_timezone(x)
}
\arguments{
\item{x}{vector with tz info, either times with it included or tz}
}
\value{
vector of OlsonName timezones if possible
}
\description{
Get timezones from various formats
}
\keyword{internal}
