% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_table.R
\name{fetch_from_table}
\alias{fetch_from_table}
\title{Fetches timeseries from a table of arguments}
\usage{
fetch_from_table(request_table, clean = TRUE)
}
\arguments{
\item{request_table}{data.frame with columns c('portal', 'gauge', 'start_time', 'end_time', 'variable', 'units', 'timeunit', 'statistic', 'datatype')}

\item{clean}{whether to clean the output to a standard set of columns, or leave as-returned}
}
\value{
dataframe of timeseries
}
\description{
Fetches timeseries from a table of arguments
}
