% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_timeseries.R
\name{fetch_timeseries}
\alias{fetch_timeseries}
\title{Fetch timeseries across formats}
\usage{
fetch_timeseries(
  portal,
  gauge,
  start_time,
  end_time,
  variable,
  units,
  timeunit,
  statistic,
  datatype,
  portal_type = "auto",
  gauge_portal = "auto",
  request_timezone = "db_default",
  return_timezone = "UTC",
  robustness = "robust",
  check_output = TRUE
)
}
\arguments{
\item{portal}{URL to Kisters KiWIS database.}

\item{gauge}{character vector of gauge numbers, as \code{station_no} for Kiwis functions (\code{site_list} for Hydstra)}

\item{start_time}{character or date or date time for the start \emph{in database default timezone}. Default NULL.}

\item{end_time}{character or date or date time for the end \emph{in database default timezone}. Default NULL.}

\item{variable}{character vector of variables we want to extract. Matches on \code{parametertype_name}}

\item{units}{units of the variable, used when there may be > 1 e.g. cumecs, ML/d for discharge. If NULL, gets all available. Matches to \code{ts_unitsymbol}}

\item{timeunit}{The time interval to request, e.g. "Daily", the default. Main values seem to be 'Daily', 'Monthly', 'Yearly', and 'AsStored' (the raw data). Matches to part of \code{ts_name}}

\item{statistic}{The aggregation statistic, e.g. "Mean", the default. Main values seem to be 'Mean', 'Max', 'Min', 'Total', though not all are available for each variable- rainfall tends to use Total, while discharge tends to use mean, max, min. Matches to part of \code{ts_name}}

\item{datatype}{as in \code{\link[=fetch_kiwis_timeseries]{fetch_kiwis_timeseries()}}, or the \code{datasource} argument for hydstra, e.g. \code{\link[=fetch_hydstra_timeseries]{fetch_hydstra_timeseries()}}}

\item{portal_type}{I want to either be able to pass this a vector of length \code{portal} or 'auto'}

\item{gauge_portal}{I want to either be able to pass this a named list of length \code{portal} or a tibble with \code{portal} and \code{gauge} columns, or 'auto' to do our best at auto-detecting.}

\item{request_timezone}{ignored if start_time and end_time are time objects, otherwise a timezone from \code{\link[=OlsonNames]{OlsonNames()}} or 'db_default'}

\item{return_timezone}{character in \code{\link[=OlsonNames]{OlsonNames()}} or one of three special cases: \code{'db_default'}, \code{'char'} or \code{'raw'}. Default 'UTC'. If 'db_default', uses the API default. BOM defaults to +10. If \code{'char'} or \code{'raw'}, returns the time column as-is from the API (A string in the format \code{'YYYY-MM-DDTHH:MM:SS+TZ'})}

\item{check_output}{logical, default TRUE- inform if there are duplicated gauges across portals or missing gauges}
}
\value{
timeseries dataframe
}
\description{
Experimental, accepts single variable for now (one type of timeseries).
Handles gauges and portals in different ways, which are not fully developed.
Mostly a lot of translating between hydstra and kiwis to allow symmetric specification.
}
