% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ts_id.R
\name{find_ts_id}
\alias{find_ts_id}
\title{Helper to extract ts_id for desired variables from \code{\link[=getTimeseriesList]{getTimeseriesList()}}}
\usage{
find_ts_id(
  portal,
  gauge,
  variable = "discharge",
  units = "ML/d",
  timeunit = "Daily",
  statistic = "Mean",
  datatype = "QaQc",
  namefilters = NULL,
  return_timezone = "char"
)
}
\arguments{
\item{portal}{URL to Kisters KiWIS database.}

\item{gauge}{character vector of gauge numbers, as \code{station_no} for Kiwis functions (\code{site_list} for Hydstra)}

\item{variable}{character vector of variables we want to extract. Matches on \code{parametertype_name}}

\item{units}{units of the variable, used when there may be > 1 e.g. cumecs, ML/d for discharge. If NULL, gets all available. Matches to \code{ts_unitsymbol}}

\item{timeunit}{The time interval to request, e.g. "Daily", the default. Main values seem to be 'Daily', 'Monthly', 'Yearly', and 'AsStored' (the raw data). Matches to part of \code{ts_name}}

\item{statistic}{The aggregation statistic, e.g. "Mean", the default. Main values seem to be 'Mean', 'Max', 'Min', 'Total', though not all are available for each variable- rainfall tends to use Total, while discharge tends to use mean, max, min. Matches to part of \code{ts_name}}

\item{datatype}{The type of data to return, default 'QaQc'. Some other options seem to be 'Recieved', 'Harmonised', and 'Obs'. \emph{Note}- 'QaQc' matches to both 'DMQaQc' and 'PR01QaQc'. In many cases only one is available, but if you get 2x too much data, check and specify which you want. Matches to part of \code{ts_name}}

\item{namefilters}{character vector giving the ability to match to other parts of \code{ts_name} in case those specified in timeunit, statistic, and datatype aren't sufficient to find the desired \code{ts_id}. One frequent occurrence is two Daily datasets that differ in whether they split at 9am or midnight, in which case you should use either \verb{namefilters = '09HR} or \verb{namefilters = '24HR}. In some situations, this can be easier than using regex, e.g.  \code{datatype = 'QaQc.*09'}}

\item{return_timezone}{character in \code{\link[=OlsonNames]{OlsonNames()}} or one of three special cases: \code{'db_default'}, \code{'char'} or \code{'raw'}. Default 'UTC'. If 'db_default', uses the API default. BOM defaults to +10. If \code{'char'} or \code{'raw'}, returns the time column as-is from the API (A string in the format \code{'YYYY-MM-DDTHH:MM:SS+TZ'})}
}
\value{
a tibble of the matching ts_ids, along with other columns
}
\description{
Helper to extract ts_id for desired variables from \code{\link[=getTimeseriesList]{getTimeseriesList()}}
}
