% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_times.R
\name{fix_times}
\alias{fix_times}
\title{Creates 14-digit character string needed for times in the API}
\usage{
fix_times(usertime, type = "hydstra")
}
\arguments{
\item{usertime}{time value from user. may be date, numeric, or character}
}
\value{
14-digit character vector
}
\description{
Creates 14-digit character string needed for times in the API
}
