% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_times.R
\name{format_chartimes}
\alias{format_chartimes}
\title{turns 14-character time string into 'YYYY-MM-DDTHH:MM:SS'}
\usage{
format_chartimes(t14)
}
\arguments{
\item{t14}{}
}
\description{
turns 14-character time string into 'YYYY-MM-DDTHH:MM:SS'
}
