% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasources_by_site.R
\name{get_datasources_by_site}
\alias{get_datasources_by_site}
\title{Gets the available datasources for each site}
\usage{
get_datasources_by_site(portal, site_list)
}
\arguments{
\item{portal}{character for the data portal (case insensitive). Default 'victoria'}

\item{site_list}{character site code, either a single site code \code{"sitenumber"}, comma-separated codes in a single string \verb{"sitenumber1, sitenumber2}, or a vector of site codes \code{c("sitenumber1", "sitenumber2")}}
}
\value{
a tibble of the sites and datasources present
}
\description{
As far as I can tell, these will be "A", "TELEM", or "TELEMCOPY", but haven't
checked all sites
}
\examples{
checkdata <- get_datasources_by_site(site_list = "233217, 405328, 405331, 405837")
}
