% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_info.R
\name{get_db_info}
\alias{get_db_info}
\title{Access to the get_db_info API call.}
\usage{
get_db_info(portal, table_name = "site", station_filter, return_type = "hash")
}
\arguments{
\item{portal}{url as elsewhere}

\item{table_name}{which table}

\item{station_filter}{a query somehow}

\item{return_type}{format of return?}
}
\value{
a dataframe
}
\description{
In development, do not use.
}
