% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{The actual API call}
\usage{
get_response(baseURL, paramlist)
}
\arguments{
\item{baseURL}{character URL that the request gets appended to}

\item{params}{a list with arguments for a particular call}
}
\value{
the response body as a list
}
\description{
The other functions build the param list and process output. This always goes
in the middle to actually hit the API. Exposing it for ad-hoc testing, etc.
}
