% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Returns the base of the request URL given state name}
\usage{
get_url(state)
}
\arguments{
\item{state}{character, name of state. Currently only accepts Victoria, New
South Wales, and Queensland (and various abbreviations, e.g. NSW, QLD, Vic)}
}
\value{
character URL for the API request
}
\description{
All the functions have been tested most with Victoria- the other states work
though, so I've built this to allow querying them too.
}
