% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_list.R
\name{get_variable_list}
\alias{get_variable_list}
\title{Gets the variables for sites and datasources}
\usage{
get_variable_list(
  baseURL = "https://data.water.vic.gov.au/cgi/webservice.exe?",
  site_list,
  datasource
)
}
\arguments{
\item{baseURL}{character, url for the base API call. Assumes Victoria, but other states may work as well, though are untested.}
}
\value{
a tibble of the variables for each site and datasource
}
\description{
Gets the variables for sites and datasources
}
\examples{
v2 <- get_variable_list(site_list = "233217, 405328, 405331, 405837",
 datasource = c('A', 'TELEM'))
}
