% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupList.R
\name{getGroupList}
\alias{getGroupList}
\title{Gets groups in the database}
\usage{
getGroupList(portal)
}
\arguments{
\item{portal}{URL to Kisters KiWIS database. Default is Australian BOM,
www.bom.gov.au/waterdata/services, but likely works for other KiWIS}
}
\value{
a tibble
}
\description{
Closest Hydstra is \code{\link[=get_groups]{get_groups()}}, neither seem particularly useful.
Might make \code{\link[=getStationList]{getStationList()}}, \code{\link[=getTimeseriesList]{getTimeseriesList()}} and \code{\link[=getTimeseriesValues]{getTimeseriesValues()}} easier, though not obviously by much.
Note that 'the group identifiers are returned as group_id but specified in getStationList, getTimeseriesList, getTimeseriesValues, and getTimeseriesValueLayer as stationgroup_id, parametergroup_id or timeseriesgroup_id' per \href{https://timeseriesdoc.sepa.org.uk/api-documentation/api-function-reference/principal-query-functions#getGroupList}{docs}
}
