% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameterList.R
\name{getParameterList}
\alias{getParameterList}
\title{Interface to the getParameterList function}
\usage{
getParameterList(
  portal,
  station_no = NULL,
  returnfields = "all",
  extra_list = list(NULL)
)
}
\arguments{
\item{portal}{URL to Kisters KiWIS database.}

\item{station_no}{gauge numbers, as \code{site_list} in the Hydstra functions.
There are many other fields that can be used to filter and select records,
but this seems to be most common and so we give it special treatment. For
others, use see \code{extra_list}}

\item{returnfields}{default 'default', otherwise 'all' to get everything
available, or comma-separated string of fields to return. \emph{Important}-
'coverage' is required to return start and end dates. It is included by
default. Full list for each function available from \href{https://timeseries.sepa.org.uk/KiWIS/KiWIS?datasource=0&service=kisters&type=queryServices&request=getrequestinfo}{Kisters docs}}

\item{extra_list}{a named list of other fields to select on. Names (usually)
should be in \code{returnfields} (or returned when \code{returnfields = 'all'}),
though not all work- see the queryfields in the \href{https://timeseries.sepa.org.uk/KiWIS/KiWIS?datasource=0&service=kisters&type=queryServices&request=getrequestinfo}{Kisters docs}.
Values should be comma-separated characters, and can contain grep wildcards
e.g. \verb{extra_list = list(station_name = 'RIVER MURRAY*)}. Can also use
groups from \code{\link[=getGroupList]{getGroupList()}}, e.g. \code{extra_list = list(stationgroup_id = '20017550')} gets the MDB_WIP_Watercourse stations.}
}
\value{
a tibble
}
\description{
It is unclear why this exists, it returns a subset of \code{\link[=getStationList]{getStationList()}}.
}
