% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationList.R
\name{getStationList}
\alias{getStationList}
\title{getStationList API call to Kisters KiWIS API}
\usage{
getStationList(
  portal,
  site_list = NULL,
  returnfields = "all",
  extra_list = list(NULL)
)
}
\arguments{
\item{site_list}{gauge numbers, as in all other functions. Converted to
\code{station_no} internally, since that is what BOM uses.}

\item{returnfields}{default 'all', otherwise comma-separated string of fields
to return}

\item{extra_list}{a named list of other fields to select on. Names should be
in \code{returnfields} (or returned when \code{returnfields = 'all'}), values should
be comma-separated characters, and can contain grep wildcards e.g.
\verb{extra_list = list(station_name = 'RIVER MURRAY*)}}

\item{baseURL}{URL to Kisters KiWIS database. Default is Australian BOM,
www.bom.gov.au/waterdata/services, but likely works for other KiWIS}
}
\description{
This is the API used by Australian Bureau of Meteorology and many others. For
consistency with similar state functions using Kisters hydllp, I have kept
the site_list argument with the same name. Any of the return fields can be
searched though, using extra_list. The equivalent state (hydllp) function is
\code{\link[=get_db_info]{get_db_info()}} (sort of).
}
