% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeseriesList.R
\name{getTimeseriesList}
\alias{getTimeseriesList}
\title{Get the list of timeseries available}
\usage{
getTimeseriesList(
  portal,
  site_list = NULL,
  returnfields = "all",
  extra_list = list(NULL)
)
}
\arguments{
\item{portal}{URL to Kisters KiWIS database. Default is Australian BOM,
www.bom.gov.au/waterdata/services, but likely works for other KiWIS}

\item{site_list}{gauge numbers, as in all other functions. Converted to
\code{station_no} internally, since that is what BOM uses.}

\item{returnfields}{default 'all', otherwise comma-separated string of fields
to return}

\item{extra_list}{a named list of other fields to select on. Names should be
in \code{returnfields} (or returned when \code{returnfields = 'all'}), values should
be comma-separated characters, and can contain grep wildcards e.g.
\verb{extra_list = list(station_name = 'RIVER MURRAY*)}}
}
\value{
A tibble of information about available timeseries
}
\description{
This is the API used by Australian Bureau of Meteorology and many others. For
consistency with similar state functions using Kisters hydllp, I have kept
the site_list argument with the same name. Any of the return fields can be
searched though, using extra_list. The equivalent state (hydllp) function is
\code{\link[=get_variable_list]{get_variable_list()}} (to a close approximation).
The available return fields (and thus factors that can be filtered) are
\verb{'station_name', 'station_no', 'station_id', 'ts_id', 'ts_name', 'ts_path', 'parametertype_id', 'parametertype_name'}, where the \code{site_list} argument matches \code{station_no} for consistency across state functions.
There is an additional \code{returnfield} option, \code{coverage}, which returns the period of record.
There may be additional undocumented returnfield options; it appears that most of what is returned by \code{getStationList} is available, e.g. 'station_latitude'
}
