% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_times.R
\name{multi_tz_check}
\alias{multi_tz_check}
\title{returns single tz from vector, with check for equality}
\usage{
multi_tz_check(tzvec)
}
\arguments{
\item{tzvec}{}
}
\description{
returns single tz from vector, with check for equality
}
