% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bom_times.R
\name{parse_bom_times}
\alias{parse_bom_times}
\title{Deal with parsing the kiwis timestamps into UTC or local time}
\usage{
parse_bom_times(timevec, timetype = "char")
}
\arguments{
\item{timevec}{output with a Timestamp column with format as BOM, e.g. \code{'2020-01-01T00:00:00.000+10:00'}}

\item{timetype}{character, one of 'char' (default), 'raw', 'UTC', or 'local'. 'char' and 'raw' both return the Timestamp as it comes from BOM, the others parse into dates.}
}
\value{
a vector, either character or POSIX
}
\description{
Local time can be fraught, if there are issues, use one of the others and do the local time parsing manually based on your particular situation.
}
\keyword{internal}
