% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_state_times.R
\name{parse_state_times}
\alias{parse_state_times}
\title{Deal with parsing the hydllp timestamps into UTC or local time}
\usage{
parse_state_times(timevec, tz, timetype)
}
\arguments{
\item{timevec}{the output of the \code{t} column from hydllp, typically a 14-digit double}

\item{tz}{a timezone vector, also returned from hydllp}

\item{timetype}{character, one of 'char' (default), 'raw', 'UTC', or 'local'. 'char' returns a character in the format \code{YYYY-MM-DDTHH:MM:SS+TZ} matching BOM-style (and so containing all necessary info). \code{'raw'} simply returns the unmodified numeric, which does not.}
}
\value{
a vector, either character, numeric, or POSIX
}
\description{
Very similar to \code{\link[=parse_bom_times]{parse_bom_times()}} but has to deal with tz differently, and return character versions differently
}
