% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasources_by_site.R
\name{plot_datasources_by_site}
\alias{plot_datasources_by_site}
\title{Plot heatmap of which sites have which datasource}
\usage{
plot_datasources_by_site(ds_by_site, returntype = "plot")
}
\arguments{
\item{ds_by_site}{tibble with \code{site} and \code{datasource} columns}

\item{returntype}{\code{'plot'} or \code{'table'} Plot much more useful- not sure why
the user wouldn't just make a table.}
}
\value{
a ggplot, typically
}
\description{
Needs work to make more general to feed it more than a few sites
}
