% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_to_gaugetime.R
\name{request_to_gaugetime}
\alias{request_to_gaugetime}
\title{Transform requested times to the gauge timezones, whatever their tz and the tz of the gauge/database}
\usage{
request_to_gaugetime(reqtime, gaugetz, request_timezone)
}
\arguments{
\item{reqtime}{requested time. Character, numeric, or POSIXt}

\item{gaugetz}{database timezone}

\item{request_timezone}{declared timezone of \code{reqtime}, ignored if \code{reqtime} is \code{POSIXt}, in which case its tz is known.}
}
\value{
requested time in the timezone \code{gaugetz}
}
\description{
Transform requested times to the gauge timezones, whatever their tz and the tz of the gauge/database
}
\keyword{internal}
