% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_table.R
\name{send_to_type}
\alias{send_to_type}
\title{Make a common request format to either kiwis or hydstra}
\usage{
send_to_type(request_table, clean = TRUE)
}
\arguments{
\item{request_table}{data.frame with columns c('portal', 'gauge',
'start_time', 'end_time', 'variable', 'units', 'timeunit', 'statistic',
'datatype')}

\item{clean}{whether to clean the output to a standard set of columns, or
leave as-returned}
}
\value{
dataframe of timeseries
}
\description{
Make a common request format to either kiwis or hydstra
}
