% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_error_catch.R
\name{ts_error_catch}
\alias{ts_error_catch}
\title{Catch API errors that don't appear until the list is unpacked}
\usage{
ts_error_catch(bodytib, call = rlang::caller_env(), .errorhandling = "stop")
}
\arguments{
\item{bodytib}{dataframe constructed from API list with \code{error_num} column}

\item{call}{see \code{\link[rlang:abort]{rlang::abort()}}; passes calling environment for clearer
messages}

\item{.errorhandling}{as in \code{\link[foreach:foreach]{foreach::foreach()}}- passed in here to handle
simply \code{abort}ing vs doing something with the error and continuing}
}
\value{
typically \code{bodytib}, unless there are errors, then determined by
\code{.errorhandling}
}
\description{
Catch API errors that don't appear until the list is unpacked
}
\keyword{internal}
