% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeseriesValues.R
\name{clean_bom_timeseries}
\alias{clean_bom_timeseries}
\title{Clean the raw output from the BOM getTimeseriesValue call}
\usage{
clean_bom_timeseries(x)
}
\arguments{
\item{x}{the response list}
}
\value{
a tibble
}
\description{
This takes a single list, and so if multiple ts_ids have been extracted, should be looped over, e.g. with \code{\link[purrr:map]{purrr::map()}}.
}
