% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts_traces.R
\name{clean_trace_list}
\alias{clean_trace_list}
\title{Cleans ts_trace API list body into tibble}
\usage{
clean_trace_list(
  responsebody,
  data_type,
  gauge = NA,
  return_timezone = "UTC",
  .errorhandling = "stop"
)
}
\arguments{
\item{responsebody}{response body from the API call to get_ts_traces}

\item{data_type}{the data_type used to calculate the statistic over the
\code{interval}, glued on for record}

\item{gauge}{character gauge name- allows building an informative error-handled output}

\item{return_timezone}{character in \code{\link[=OlsonNames]{OlsonNames()}} or one of three special cases: \code{'db_default'}, \code{'char'} or \code{'raw'}. Default 'UTC'. If 'db_default', uses the API default. If \code{'char'}, returns a string in the format \code{'YYYY-MM-DDTHH:MM:SS+TZ'}, having all needed info (and matching Kiwis returns). If \code{'raw'}, returns the time column as-is from the API (a 14-digit string of numbers 'YYYMMDDHHMMSS')
\itemize{
\item \code{"sxvlist"} returns a list with an separate tibble for each site x variable combination
}}

\item{.errorhandling}{as in \code{\link[foreach:foreach]{foreach::foreach()}} (but handled in
\code{\link[=api_error_catch]{api_error_catch()}}) Default 'stop'. Made available here primarily to use
'pass' so big requests don't die due to API errors. \strong{Be careful}- those
errors then just get passed and so the data will be missing. Only currently
implemented and working in \code{\link[=get_ts_traces2]{get_ts_traces2()}}}
}
\value{
a tibble with the rectangled response
}
\description{
Cleans ts_trace API list body into tibble
}
