% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts_traces.R
\name{clean_trace_list}
\alias{clean_trace_list}
\title{Cleans ts_trace API list body into tibble}
\usage{
clean_trace_list(
  responsebody,
  data_type,
  gauge = NA,
  timetype = "char",
  .errorhandling = "stop"
)
}
\arguments{
\item{responsebody}{response body from the API call to get_ts_traces}

\item{data_type}{the data_type used to calculate the statistic over the
\code{interval}, glued on for record}

\item{gauge}{character gauge name- allows building an informative error-handled output}

\item{.errorhandling}{as in \code{\link[foreach:foreach]{foreach::foreach()}} (but handled in
\code{\link[=api_error_catch]{api_error_catch()}}) Default 'stop'. Made available here primarily to use
'pass' so big requests don't die due to API errors. \strong{Be careful}- those
errors then just get passed and so the data will be missing. Only currently
implemented and working in \code{\link[=get_ts_traces2]{get_ts_traces2()}}}
}
\value{
a tibble with the rectangled response
}
\description{
Cleans ts_trace API list body into tibble
}
