% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{The actual API call}
\usage{
get_response(
  baseURL,
  api_body_list = NULL,
  api_query_list = list(nullquery = NULL),
  .errorhandling = "stop"
)
}
\arguments{
\item{baseURL}{character URL that the request gets appended to}

\item{api_body_list}{a list with arguments for a particular call}

\item{.errorhandling}{intended to allow passing or removing errors. Currently
only implemented in a way that works for \code{\link[=get_ts_traces2]{get_ts_traces2()}}; see
documentation there.}
}
\value{
the response body as a list
}
\description{
The other functions build the param list and process output. This always goes
in the middle to actually hit the API. Exposing it for ad-hoc testing, etc.
}
