% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasources_by_site.R
\name{get_sites_by_datasource}
\alias{get_sites_by_datasource}
\title{Gets sites that have a given datasource}
\usage{
get_sites_by_datasource(portal, datasources)
}
\arguments{
\item{portal}{character for the data portal (case insensitive). Default 'victoria'}

\item{datasources}{as in \code{\link[=get_ts_traces]{get_ts_traces()}}, but can be a vector. As far as I can tell, options are \code{'A'}, \code{'TELEM'}, \code{'TELEMCOPY'}. If multiple, \code{c('A', 'TELEM')}}
}
\value{
a tibble with sites and datasources
}
\description{
Useful for getting lists of most sites
}
\examples{
sxd <- get_sites_by_datasource(datasources = c('A', 'TELEM'))
}
