% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_list.R
\name{get_variable_list}
\alias{get_variable_list}
\title{Gets the variables for sites and datasources}
\usage{
get_variable_list(portal, site_list, datasource)
}
\arguments{
\item{portal}{character for the data portal (case insensitive). Default 'victoria'}

\item{site_list}{character site code, either a single site code \code{"sitenumber"}, comma-separated codes in a single string \verb{"sitenumber1, sitenumber2}, or a vector of site codes \code{c("sitenumber1", "sitenumber2")}}

\item{datasource}{as in \code{\link[=get_ts_traces]{get_ts_traces()}}, but can be a vector. As far as I can tell, options are \code{'A'}, \code{'TELEM'}, \code{'TELEMCOPY'}. If multiple, \code{c('A', 'TELEM')}}
}
\value{
a tibble of the variables for each site and datasource
}
\description{
Gets the variables for sites and datasources
}
\examples{
v2 <- get_variable_list(site_list = "233217, 405328, 405331, 405837",
 datasource = c('A', 'TELEM'))
}
