% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogauge-package.R
\docType{package}
\name{hydrogauge-package}
\alias{hydrogauge}
\alias{hydrogauge-package}
\title{hydrogauge: Pulls information from the Kisters Kiwis and Hydstra APIs}
\description{
This package provides R wrappers to (some) of the API calls possible to the Victorian gauge network. I have tried to provide functionality to query available options for the different arguments.
}
\author{
\strong{Maintainer}: Galen Holt \email{g.holt@deakin.edu.au} (\href{https://orcid.org/0000-0002-7455-9275}{ORCID})

}
\keyword{internal}
